write initial functionaolity to check of there is an order for a payment to posted to
